classic REPL の段階退役準備
