+++ internal/agent/agent_interactive_context_size_test.go
